refactored code for constants and split route for readiability and maintainability
