+++ app/constants/max-files.js
@@ -1,0 +1,3 @@
+module.exports = {
+  MAX_FILES: 3
+}
